+++ server/routers/galleryRouter.js
